refactor# update api test demo
