--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/apis/common/PageFormParam.java
@@ -1,18 +1,0 @@
-package org.apache.dolphinscheduler.test.apis.common;
-
-public enum PageFormParam {
-    PAGE_SIZE("pageSize"),
-    PAGE_NO("pageNo"),
-    SEARCH_VAL("searchVal");
-
-    PageFormParam(String param) {
-        this.param = param;
-
-    }
-
-    private final String param;
-
-    public String getParam() {
-        return param;
-    }
-}
