--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/apis/common/RequestMethod.java
@@ -1,16 +1,0 @@
-package org.apache.dolphinscheduler.test.apis.common;
-
-public enum RequestMethod {
-    POST("Post"),
-    PUT("put");
-
-    RequestMethod(String method) {
-        this.method = method;
-    }
-
-    private final String method;
-
-    public String getMethod() {
-        return method;
-    }
-}
