--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/apis/configCenter/tenant/TenantEndPoints.java
@@ -1,5 +1,0 @@
-package org.apache.dolphinscheduler.test.apis.configCenter.tenant;
-
-public interface TenantEndPoints {
-
-}
