--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/apis/configCenter/tenant/entity/TenantRequestEntity.java
@@ -1,9 +1,0 @@
-package org.apache.dolphinscheduler.test.apis.configCenter.tenant.entity;
-
-public class TenantRequestEntity extends TenantBaseEntity {
-    private int id;
-    public void setId(int id) {
-        this.id = id;
-    }
-
-}
