--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/apis/login/entity/LoginResponseEntity.java
@@ -1,13 +1,0 @@
-package org.apache.dolphinscheduler.test.apis.login.entity;
-
-public class LoginResponseEntity {
-    private String SessionId;
-
-    public String getSessionId() {
-        return SessionId;
-    }
-
-    public void setSessionId(String sessionId) {
-        SessionId = sessionId;
-    }
-}
