--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/base/IBaseEntity.java
@@ -1,3 +1,0 @@
-package org.apache.dolphinscheduler.test.base;
-
-public interface IBaseEntity {}
