--- dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/base/IRestResponse.java
@@ -1,19 +1,0 @@
-package org.apache.dolphinscheduler.test.base;
-
-import io.restassured.response.Response;
-
-public interface IRestResponse<T>{
-    public T getBody();
-
-    public String getContent();
-
-    public int getStatusCode();
-
-    public boolean isSuccessful();
-
-    public String getStatusDescription();
-
-    public Response getResponse();
-
-    public Exception getException();
-}
