+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/cases/ui/TenantE2ETest.java
@@ -1,25 +1,26 @@
 package org.apache.dolphinscheduler.test.cases.ui;
 
-import com.devskiller.jfairy.Fairy;
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.apache.dolphinscheduler.test.cases.common.AbstractTenantApiTest;
 import org.apache.dolphinscheduler.test.core.Browser;
-import org.apache.dolphinscheduler.test.pages.security.common.SecurityPage;
 import org.apache.dolphinscheduler.test.pages.security.tenantManage.TenantManagePage;
 import org.apache.dolphinscheduler.test.pages.login.LoginPage;
-import org.apache.dolphinscheduler.test.pages.navBar.NavBarPage;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.slf4j.Logger;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.slf4j.LoggerFactory.getLogger;
+import static org.testcontainers.shaded.org.awaitility.Awaitility.await;
 
-public class TenantE2ETest {
+public class TenantE2ETest extends AbstractTenantApiTest {
+    private static final String tenant = System.getProperty("user.name");
     private static final Logger log = getLogger(TenantE2ETest.class);
-    private final Fairy fairy = Fairy.create();
     public WebDriver driver;
     public Browser browser;
 
@@ -44,11 +45,18 @@ public class TenantE2ETest {
 
     @Test
     void testCreateTenant() throws Exception {
-        TenantManagePage tenantManagePage = browser.
+        TenantManagePage page = browser.
                 toPage(LoginPage.class).
                 loginAs().
                 toSecurityTab().
                 toTenantManage();
-        tenantManagePage.create(fairy.person().getUsername(), "eeee");
+        String tenantCode = fairy.person().getUsername();
+        page.create(tenantCode, fairy.person().getMobileTelephoneNumber());
+
+        await().untilAsserted(() -> assertThat(page.getTenantList())
+                .as("Tenant list should contain newly-created tenant")
+                .extracting(WebElement::getText)
+                .anyMatch(it -> it.contains(tenantCode)));
+
     }
 }
