+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/IEndPointFactory.java
@@ -1,0 +1,10 @@
+package org.apache.dolphinscheduler.test.endpoint;
+
+import io.restassured.specification.RequestSpecification;
+import org.apache.dolphinscheduler.test.endpoint.api.security.tenant.TenantEndPoints;
+
+public interface IEndPointFactory {
+
+    TenantEndPoints createTenantEndPoints();
+
+}
