+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/api/common/FormParam.java
@@ -1,0 +1,21 @@
+package org.apache.dolphinscheduler.test.endpoint.api.common;
+
+public enum FormParam {
+    SESSION_ID("sessionId"),
+    CODE("code"),
+    MSG("msg"),
+    DATA("data"),
+    SUCCESS("success"),
+    FAILED("failed");
+
+
+    FormParam(String param) {
+        this.param = param;
+    }
+
+    private final String param;
+
+    public String getParam() {
+        return param;
+    }
+}
