+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/api/common/RequestMethod.java
@@ -1,0 +1,16 @@
+package org.apache.dolphinscheduler.test.endpoint.api.common;
+
+public enum RequestMethod {
+    POST("Post"),
+    PUT("put");
+
+    RequestMethod(String method) {
+        this.method = method;
+    }
+
+    private final String method;
+
+    public String getMethod() {
+        return method;
+    }
+}
