+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/api/login/entity/LoginResponseEntity.java
@@ -1,0 +1,13 @@
+package org.apache.dolphinscheduler.test.endpoint.api.login.entity;
+
+public class LoginResponseEntity {
+    private String SessionId;
+
+    public String getSessionId() {
+        return SessionId;
+    }
+
+    public void setSessionId(String sessionId) {
+        SessionId = sessionId;
+    }
+}
