+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/api/security/tenant/entity/TenantRequestEntity.java
@@ -1,0 +1,9 @@
+package org.apache.dolphinscheduler.test.endpoint.api.security.tenant.entity;
+
+public class TenantRequestEntity extends TenantBaseEntity {
+    private int id;
+    public void setId(int id) {
+        this.id = id;
+    }
+
+}
