+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/base/AbstractBaseEntity.java
@@ -1,0 +1,13 @@
+package org.apache.dolphinscheduler.test.endpoint.base;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.util.Map;
+
+public abstract class AbstractBaseEntity implements IBaseEntity {
+
+    public Map<String, Object> toMap() {
+        ObjectMapper objMapper = new ObjectMapper();
+        return objMapper.convertValue(this, Map.class);
+    }
+}
