+++ dolphinscheduler-test/dolphinscheduler-test-case/src/test/java/org/apache/dolphinscheduler/test/endpoint/base/IBaseEntity.java
@@ -1,0 +1,3 @@
+package org.apache.dolphinscheduler.test.endpoint.base;
+
+public interface IBaseEntity {}
