+++ dolphinscheduler-test/dolphinscheduler-test-core/src/main/java/org/apache/dolphinscheduler/test/core/Module.java
@@ -1,4 +1,5 @@
 package org.apache.dolphinscheduler.test.core;
 
 public class Module extends Page{
+    public Module() {}
 }
